🎨 Style: JOB_CATEGORIES 변수를 constants 폴더로 이동
