+++ app/_constants/JobCategoryData.ts
@@ -1,0 +1,14 @@
+const JOB_CATEGORIES = {
+  ALL: "전체",
+  FRONTEND: "프론트엔드",
+  BACKEND: "백엔드",
+  DESIGNER: "디자이너",
+  IOS: "IOS",
+  ANDROID: "안드로이드",
+  DEVOPS: "데브옵스",
+  PLANNER: "기획자",
+  FULLSTACK: "풀스택",
+  ETC: "기타",
+};
+
+export default JOB_CATEGORIES;
